requirements.txt
amqp==5.0.5
asgiref==3.3.1
Babel==2.9.0
botocore==1.20.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.4
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-configurations==2.2
django-formtools==2.2
django-modeladmin-reorder==0.3.1
django-request-logging==0.7.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
environs==9.3.1
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.10.0
names==0.3.0
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.14
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==2.0.1
